requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2.1
django-cors-headers==3.7.0
django-oauth-toolkit==1.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jwcrypto==0.8
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.2.0
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
